pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "daoc-bot"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]

[project.scripts]
daoc-bot = "daoc_bot.__main__:bot.run"

[tool.setuptools.packages.find]
where = ["."]
include = ["daoc_bot*"]

# ── MyPy ──────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
files = ["daoc_bot"]
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true
# discord.py stubs are incomplete; silence whole-module noise
[[tool.mypy.overrides]]
module = ["discord.*"]
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "error",
    # discord.py uses deprecated asyncio APIs internally
    "ignore::DeprecationWarning:discord",
]

# ── Ruff ──────────────────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
